refactor(callback): move MonitorCallback to a standalone `callbacks.py`
